feat: move convert function to utils
